+++ src/ann/layers/loss.cpp
@@ -1,12 +1,12 @@
 #include "ann/layers/loss.h"
 
 namespace Layer {
-double Loss::mean() {
-  double outputSum{};
+float Loss::mean() {
+  float outputSum{};
 
   for (size_t i{}; i < m_output->size(); ++i)
     outputSum += (*m_output)[i];
 
-  return outputSum / static_cast<double>(m_output->size());
+  return outputSum / static_cast<float>(m_output->size());
 }
 } // namespace Layer
